AG-1752-fix-master-detail-example-for-frameworks
